# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "支持多租户、多存储后端的软件升级管理服务",
        "title": "软件升级管理系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/update/check": {
            "get": {
                "description": "客户端检查是否有新版本可用",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "升级服务"
                ],
                "summary": "检查更新",
                "parameters": [
                    {
                        "type": "string",
                        "description": "软件标识",
                        "name": "software_slug",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "当前版本号",
                        "name": "current_version",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "平台",
                        "name": "platform",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "device_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "API密钥",
                        "name": "api_key",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.CheckUpdateResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/update/download/{id}": {
            "get": {
                "description": "下载指定版本的安装包",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "升级服务"
                ],
                "summary": "下载版本包",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "版本ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API密钥",
                        "name": "api_key",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/update/download/{id}/direct": {
            "get": {
                "description": "通过服务器中转下载安装包",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "升级服务"
                ],
                "summary": "直接下载版本包",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "版本ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API密钥",
                        "name": "api_key",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/update/latest": {
            "get": {
                "description": "获取软件的最新发布版本信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "升级服务"
                ],
                "summary": "获取最新版本",
                "parameters": [
                    {
                        "type": "string",
                        "description": "软件标识",
                        "name": "software_slug",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Version"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "api.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "model.Software": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "platform": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "tenant_id": {
                    "description": "多租户预留字段",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "versions": {
                    "description": "关联",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Version"
                    }
                }
            }
        },
...
        "model.Version": {
            "type": "object",
            "properties": {
                "changelog": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_forced": {
                    "type": "boolean"
                },
                "is_incremental": {
                    "type": "boolean"
                },
                "min_version": {
                    "type": "string"
                },
                "package_hash": {
                    "type": "string"
                },
                "package_size": {
                    "type": "integer"
                },
                "package_url": {
                    "type": "string"
                },
                "published_at": {
                    "type": "string"
                },
                "software": {
                    "description": "关联",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Software"
                        }
                    ]
                },
                "software_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "tenant_id": {
                    "description": "多租户预留字段",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                },
                "version_code": {
                    "type": "integer"
                }
            }
        },
        "service.CheckUpdateResponse": {
            "type": "object",
            "properties": {
                "changelog": {
                    "type": "string"
                },
                "download_url": {
                    "type": "string"
                },
                "has_update": {
                    "type": "boolean"
                },
                "is_forced": {
                    "type": "boolean"
                },
                "is_incremental": {
                    "type": "boolean"
                },
                "min_version": {
                    "type": "string"
                },
                "package_hash": {
                    "type": "string"
                },
                "package_size": {
                    "type": "integer"
                },
                "published_at": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                },
                "version_code": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}